# Conflicts:
#	Platform/CrossPlatform/Effect.cpp
#	Platform/DirectX12/Effect.cpp
#	PlugIns/Unity/Project5/Assets/Simul/Presets/Cirrocumulus.asset
#	PlugIns/Unity/Project5/Assets/Simul/Presets/CirrocumulusKeyframe.asset
#	PlugIns/Unity/Project5/Assets/Simul/Presets/Cirrus.asset
#	PlugIns/Unity/Project5/Assets/Simul/Presets/Cumulus.asset
#	PlugIns/Unity/Project5/Assets/Simul/Presets/CumulusKeyframe.asset
#	PlugIns/Unity/Project5/Assets/Simul/Presets/NoCloudsKeyframe.asset
#	PlugIns/Unity/Project5/Assets/Simul/Presets/Stratocumulus.asset
#	PlugIns/Unity/Project5/Assets/Simul/Presets/StratocumulusKeyframe.asset
#	PlugIns/Unity/Project5/Assets/Simul/Presets/Stratus.asset
#	PlugIns/Unity/Project5/Assets/Simul/demo/New Terrain.asset
#	PlugIns/Unity/Project5/Assets/Simul/demo/demo.asset

+++ Platform/CrossPlatform/CommandLineParams.h
@@ -87,7 +87,7 @@ namespace simul
 			for(int i=0;i<argCount;i++)
 			{
 				std::string str=base::WStringToString(szArgList[i]);
-				int len=str.length();
+				int len=(int)str.length();
 				args[i]=new char[len+1];
 				strcpy_s(args[i],len+1,str.c_str());
 				args[i][str.length()]=0;
